refactor(InteractionDataOptionAttachment): change value to snowflake
