+++ deno/payloads/v9/_interactions/_applicationCommands/chatInput.ts
@@ -1,6 +1,5 @@
-import type { APIRole, APIUser, ChannelType } from '../../mod.ts';
+import type { APIRole, APIUser, ChannelType, APIAttachment } from '../../mod.ts';
 import type { Snowflake } from '../../../../globals.ts';
-import type { APIAttachment } from '../../channel.ts';
 import type { APIDMInteractionWrapper, APIGuildInteractionWrapper } from '../base.ts';
 import type {
 	APIApplicationCommandInteractionWrapper,
@@ -212,7 +211,7 @@ export interface ApplicationCommandInteractionDataOptionInteger
 
 export type ApplicationCommandInteractionDataOptionAttachment = InteractionDataOptionBase<
 	ApplicationCommandOptionType.Attachment,
-	APIAttachment
+	Snowflake
 >;
 
 export interface ApplicationCommandInteractionDataOptionNumber
